feat:实现接入层

+++ mpp-support/src/main/java/com/boat/mpp/support/dao/MessageTemplateDao.java
@@ -1,9 +1,8 @@
 package com.boat.mpp.support.dao;
 
-import com.boat.mpp.support.entity.MessageTemplate;
+import com.boat.mpp.support.domain.MessageTemplate;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
@@ -15,7 +14,7 @@ public interface MessageTemplateDao extends JpaRepository<MessageTemplate, Long>
      * @param pageable 分页对象
      * @return 短信模板
      */
-    List<MessageTemplate> findAllByIsDeletedEquals(Byte isDeleted, Pageable pageable);
+    List<MessageTemplate> findAllByIsDeletedEquals(Integer isDeleted, Pageable pageable);
 
 
     /**
@@ -23,5 +22,5 @@ public interface MessageTemplateDao extends JpaRepository<MessageTemplate, Long>
      * @param isDeleted 0：未删除 1：删除
      * @return 未删除的条数
      */
-    Long countByIsDeletedEquals(Byte isDeleted);
+    Long countByIsDeletedEquals(Integer isDeleted);
 }
